+++ cmd/experiments/web-ui/handlers.go
@@ -8,6 +8,7 @@ import (
 	"strings"
 	"time"
 
+	"github.com/go-go-golems/geppetto/pkg/conversation"
 	"github.com/google/uuid"
 )
 
@@ -72,7 +73,7 @@ func RegisterHandlers(server *Server, tmpl *template.Template) {
 
 		if !exists {
 			server.logger.Info().Str("client_id", clientID).Msg("Creating new client")
-			client = NewSSEClient(clientID)
+			client = NewSSEClient(clientID, tmpl)
 			server.RegisterClient(client)
 		}
 		defer server.UnregisterClient(client.ID)
@@ -107,48 +108,6 @@ func RegisterHandlers(server *Server, tmpl *template.Template) {
 		}
 	})
 
-	// Handle start endpoint - creates and starts a new step
-	http.HandleFunc("/start", func(w http.ResponseWriter, r *http.Request) {
-		// Create new client ID
-		clientID := uuid.New().String()
-		server.logger.Info().Str("client_id", clientID).Msg("Starting new chat")
-
-		// Create new client
-		client := NewSSEClient(clientID)
-		server.RegisterClient(client)
-
-		if err := server.CreateStep(clientID); err != nil {
-			server.logger.Error().Err(err).Str("client_id", clientID).Msg("Failed to create step")
-			http.Error(w, fmt.Sprintf("Error creating step: %v", err), http.StatusInternalServerError)
-			return
-		}
-
-		// XXX we need to properly handle the context here, at a client level
-		err := server.router.RunHandlers(context.Background())
-		if err != nil {
-			server.logger.Error().Err(err).Str("client_id", clientID).Msg("Failed to run handlers")
-			http.Error(w, fmt.Sprintf("Error running handlers: %v", err), http.StatusInternalServerError)
-			return
-		}
-
-		if err := server.StartStep(clientID); err != nil {
-			server.logger.Error().Err(err).Str("client_id", clientID).Msg("Failed to start step")
-			http.Error(w, fmt.Sprintf("Error starting step: %v", err), http.StatusInternalServerError)
-			return
-		}
-
-		// Return the chat container template with the new client ID
-		data := TemplateData{
-			ClientID: clientID,
-		}
-		w.Header().Set("HX-Push-Url", fmt.Sprintf("/?client_id=%s", clientID))
-		if err := tmpl.ExecuteTemplate(w, "chat-container", data); err != nil {
-			server.logger.Error().Err(err).Msg("Failed to render chat container")
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-			return
-		}
-	})
-
 	// Handle chat endpoint - processes chat messages
 	http.HandleFunc("/chat", func(w http.ResponseWriter, r *http.Request) {
 		if r.Method != http.MethodPost {
@@ -159,18 +118,24 @@ func RegisterHandlers(server *Server, tmpl *template.Template) {
 		clientID := r.FormValue("client_id")
 		message := r.FormValue("message")
 
-		if clientID == "" || message == "" {
-			http.Error(w, "Missing client_id or message", http.StatusBadRequest)
-			return
+		// For new chats, generate a client ID
+		if clientID == "" {
+			clientID = uuid.New().String()
+			server.logger.Info().Str("client_id", clientID).Msg("Starting new chat")
 		}
 
-		server.logger.Info().
-			Str("client_id", clientID).
-			Str("message", message).
-			Msg("Received chat message")
+		// Get or create client
+		server.clientsMux.RLock()
+		client, exists := server.clients[clientID]
+		server.clientsMux.RUnlock()
+
+		if !exists {
+			client = NewSSEClient(clientID, tmpl)
+			server.RegisterClient(client)
+		}
 
 		// Create new step for this message
-		if err := server.CreateStep(clientID); err != nil {
+		if err := client.CreateStep(server.router); err != nil {
 			server.logger.Error().Err(err).Str("client_id", clientID).Msg("Failed to create step")
 			http.Error(w, fmt.Sprintf("Error creating step: %v", err), http.StatusInternalServerError)
 			return
@@ -184,14 +149,40 @@ func RegisterHandlers(server *Server, tmpl *template.Template) {
 			return
 		}
 
-		// Start step
-		if err := server.StartStep(clientID); err != nil {
+		// Create conversation messages
+		msgs := []*conversation.Message{
+			conversation.NewChatMessage(conversation.RoleSystem, "You are a helpful assistant."),
+		}
+
+		// Add user message if provided
+		if message != "" {
+			msgs = append(msgs, conversation.NewChatMessage(conversation.RoleUser, message))
+		} else {
+			msgs = append(msgs, conversation.NewChatMessage(conversation.RoleUser, "Hello! Please tell me a short story about a robot."))
+		}
+
+		// Start step with conversation
+		if err := client.StartStep(context.Background(), msgs); err != nil {
 			server.logger.Error().Err(err).Str("client_id", clientID).Msg("Failed to start step")
 			http.Error(w, fmt.Sprintf("Error starting step: %v", err), http.StatusInternalServerError)
 			return
 		}
 
-		// Return empty response as events will be sent via SSE
+		// For new chats, return the chat container template
+		if !exists {
+			data := TemplateData{
+				ClientID: clientID,
+			}
+			w.Header().Set("HX-Push-Url", fmt.Sprintf("/?client_id=%s", clientID))
+			if err := tmpl.ExecuteTemplate(w, "chat-container", data); err != nil {
+				server.logger.Error().Err(err).Msg("Failed to render chat container")
+				http.Error(w, err.Error(), http.StatusInternalServerError)
+				return
+			}
+			return
+		}
+
+		// For existing chats, return empty response as events will be sent via SSE
 		w.WriteHeader(http.StatusOK)
 	})
 }
