+++ cmd/experiments/web-ui/types.go
@@ -2,12 +2,8 @@ package main
 
 import (
 	"context"
-	"html/template"
-	"sync"
 
-	"github.com/go-go-golems/geppetto/pkg/events"
 	"github.com/go-go-golems/geppetto/pkg/steps/ai/chat"
-	"github.com/rs/zerolog"
 )
 
 // TemplateData holds data for rendering templates
@@ -20,18 +16,4 @@ type StepInstance struct {
 	Step   *chat.EchoStep
 	Topic  string
 	Cancel context.CancelFunc
-}
-
-// Server is the main server struct that handles all web UI functionality
-type Server struct {
-	tmpl       *template.Template
-	router     *events.EventRouter
-	clients    map[string]*SSEClient
-	steps      map[string]*StepInstance
-	clientsMux sync.RWMutex
-	stepsMux   sync.RWMutex
-	logger     zerolog.Logger
-	metrics    struct {
-		TotalDroppedMsgs int64
-	}
 }
